Moved plots to BaseTimeseries
